+++ .github/workflows/lint-test-code.yml
@@ -4,8 +4,6 @@ name: Linting, Testing and deployment
 on:
   push:
     branches: [ dev ]
-  pull_request:
-    branches: [ main ]
 
 jobs:
   build:
@@ -25,7 +23,7 @@ jobs:
     - name: Install dev dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install fastapi flake8 pytest autopep8
+        python -m pip install requests fastapi flake8 pytest autopep8
     - name: Lint with autopep8 and check with flake8
       run: |
         # lint all python files in project
