+++ .github/workflows/test-build-deploy-image.yml
@@ -19,7 +19,7 @@ jobs:
           python-version: "3.8"
 
       - name: Install Dependencies
-        run: python -m pip install docker pytest
+        run: python -m pip install docker pytest fastapi six
 
       - name: Login to GitHub Container Registry
         uses: docker/login-action@v1
@@ -34,11 +34,11 @@ jobs:
         with:
           context: .
           push: false
-          tags: ghcr.io/user/app:latest
+          tags: ghcr.io/ismael-su/python-ci:latest
 
       - name: Test Image
         run: pytest tests/image/
 
       - name: Push
-        run: docker push ghcr.io/user/app:latest
+        run: docker push ghcr.io/ismael-su/python-ci:latest
